Function cs220::assignments::assignment03::custom_operators::my_option_op_or

pub fn my_option_op_or<T, F: FnOnce(T, T) -> T>(
    v1: MyOption<T>,
    v2: MyOption<T>,
    f: F,
) -> MyOption<T>

Custom operator: option_op_or(v1, v2, f). If neither v1 nor v2 is Some, returns None. If exactly one is Some, returns the same Some value. If both are Some, apply the values inside Some to f and wrap the resulting value inside Some.

Examples

use cs220::assignments::assignment03::custom_operators::*;
fn product(a: i32, b: i32) -> i32 {
    a * b
}

assert_eq!(my_option_op_or(MyOption::MyNone, MyOption::MyNone, product), MyOption::MyNone);
assert_eq!(my_option_op_or(MyOption::MySome(3), MyOption::MyNone, product), MyOption::MySome(3));
assert_eq!(my_option_op_or(MyOption::MySome(3), MyOption::MySome(5), product), MyOption::MySome(15));