Module cs220::assignments::assignment03::small_exercises

Small problems.

Enums

DayOfWeek

Day of week.

TypeEvent

Events in a text editor.

Functions

median

Given a list of integers, returns its median (when sorted, the value in the middle position).

mode

Given a list of integers, returns its smallest mode (the value that occurs most often; a hash map will be helpful here).

next_weekday

The next day of week.

organize

Converts HR commands to the organization table.

piglatin

Converts the given string to Pig Latin. Use the rules below to translate normal English into Pig Latin.

use_editor

Starting from an empty string and an empty clipboard, processes the given events in order and returns the resulting string.