Function cs220::assignments::assignment03::small_exercises::median

pub fn median(values: Vec<isize>) -> Option<isize>

Given a list of integers, returns its median (when sorted, the value in the middle position).

For a data set x of n elements, the median can be defined as follows:

If n is odd, the median is (n+1)/2-th smallest element of x.

If n is even, the median is (n/2)+1-th smallest element of x.

For example, the following list of seven numbers,

vec![1, 3, 3, 6, 7, 8, 9]

has the median of 6, which is the fourth value. And for this data set of eight numbers,

vec![1, 2, 3, 4, 5, 6, 8, 9]

it has the median of 5, which is the fifth value.

Returns None if the list is empty.