Function cs220::assignments::assignment03::small_exercises::organize

pub fn organize(commands: Vec<String>) -> HashMap<String, HashSet<String>>

Converts HR commands to the organization table.

If the commands are as follows:

vec!["Add Amir to Engineering", "Add Sally to Sales", "Remove Jeehoon from Sales", "Move Amir from Engineering to Sales"]

The return value should be:

["Sales" -> ["Amir", "Sally"]]

The result is a map from department to the list of its employees.

An empty department should not appear in the result.

There are three commands: “Add {person} to {department}”, “Remove {person} from {department}”, and “Move {person} from {department} to {department}”.

If a command is not executable, then it’s ignored.

There is no space in the name of the person and department.

See the test function for more details.