Module cs220::assignments::assignment08::church

Church Numerals

This exercise involves the use of “Church numerals”, a representation of natural numbers using lambda calculus, named after Alonzo Church. Each Church numeral corresponds to a natural number n and is represented as a higher-order function that applies a given function f n times.

For more information, see:

https://en.wikipedia.org/wiki/Church_encoding

https://opendsa-server.cs.vt.edu/OpenDSA/Books/PL/html/ChurchNumerals.html

Functions

add

Implement a function to add two Church numerals.

exp

Implement a function to raise one Church numeral to the power of another. This is the Church numeral equivalent of the natural number operation of exponentiation. Given two natural numbers n and m, the function should return a Church numeral that represents n to the power of m. The key is to convert n and m to Church numerals, and then apply the Church numeral for m (the exponent) to the Church numeral for n (the base). Note: This function should be implemented WITHOUT using the to_usize or any pow-like method.

from_usize

Implement a function to convert a usize type to a Church numeral.

mult

Implement a function to multiply (mult) two Church numerals.

one

This function returns a Church numeral equivalent of the natural number 1. It takes a function f and applies it exactly once.

succ

Implement a function to add 1 to a given Church numeral.

to_usize

Implement a function to convert a Church numeral to a usize type.

two

This function returns a Church numeral equivalent of the natural number 2. It takes a function f and applies it twice.

zero

This function represents the Church numeral for zero. As zero applications of f should leave the argument unchanged, the function simply returns the input.

Type Definitions

Church

Church numerals are represented as higher-order functions that take a function f