Function cs220::assignments::assignment08::church::exp

pub fn exp<T: 'static>(n: usize, m: usize) -> Church<T>

Implement a function to raise one Church numeral to the power of another. This is the Church numeral equivalent of the natural number operation of exponentiation. Given two natural numbers n and m, the function should return a Church numeral that represents n to the power of m. The key is to convert n and m to Church numerals, and then apply the Church numeral for m (the exponent) to the Church numeral for n (the base). Note: This function should be implemented WITHOUT using the to_usize or any pow-like method.