Struct cs220::assignments::assignment09::bigint::BigInt

pub struct BigInt {
    pub carrier: Vec<u32>,
}

An signed integer with infinite precision implemented with an “carrier” vector of u32s.

The vector is interpreted as a base 2^(32 * (len(carrier) - 1)) integer, where negative integers are represented in their 2’s complement form.

For example, the vector vec![44,345,3] represents the integer 44 * (2^32)^2 + 345 * (2^32) + 3, and the vector vec![u32::MAX - 5, u32::MAX - 7] represents the integer - (5 * 2^32 + 8)

You will implement the Add and Sub trait for this type.

Unlike standard fix-sized intergers in Rust where overflow will panic, the carrier is extended to save the overflowed bit. On the contrary, if the precision is too much (e.g, vec![0,0] is used to represent 0, where vec![0] is sufficent), the carrier is truncated.

See this section for a rouge guide on implementation, while keeping in mind that the carrier should be extended to deal with overflow.

The sign_extension(), two_complement(), and truncate() are non-mandatory helper methods.

For testing and debugging pruposes, the Display trait is implemented for you, which shows the integer in hexadecimal form.

Fields
carrier: Vec<u32>

The carrier for BigInt.

Note that the carrier should always be non-empty.

Implementations

impl BigInt

pub fn new(n: u32) -> Self

Create a new BigInt from a usize.

pub fn new_large(carrier: Vec<u32>) -> Self

Creates a new BigInt from a Vec<u32>.

Panic

Panics if carrier is empty.