Function cs220::assignments::assignment09::small_exercises::is_fibonacci

pub fn is_fibonacci(inner: impl Iterator<Item = i64>) -> bool

Returns whether the given sequence is a fibonacci sequence starts from the given sequence’s first two terms.

Returns true if the length of sequence is less or equal than 2.

Exmample

use cs220::assignments::assignment09::small_exercises::*;

assert_eq!(is_fibonacci([1, 1, 2, 3, 5, 8, 13].into_iter()), true);
assert_eq!(is_fibonacci([1, 1, 2, 3, 5, 8, 14].into_iter()), false);