Function cs220::assignments::assignment09::small_exercises::position_median

pub fn position_median<T: Ord>(inner: Vec<T>) -> Option<usize>

Return the position of the median element in the vector.

For a data set x of n elements, the median can be defined as follows:

If n is odd, the median is (n+1)/2-th smallest element of x.

If n is even, the median is (n/2)+1-th smallest element of x.

Please following these rules:

If the list is empty, returns None.

If several elements are equally median, the position of the first of them is returned.

Exmaple

use cs220::assignments::assignment09::position_median;

assert_eq!(position_median(vec![1, 3, 3, 6, 7, 8, 9]), Some(3));
assert_eq!(position_median(vec![1, 3, 3, 3]), Some(1));