Module cs220::assignments::assignment10::small_exercises

Small exercises.

Enums

File

File

Node

Represents a node of tree data structure.

Functions

du_sort

Given a file, summarize all subfiles and sizes in ascending order of size.

inversion

Returns the pairs of (i, j) where i < j and inner[i] > inner[j] in increasing order.

natural_join

Returns the natural join of two tables using the first column as the join argument. That is, for each pair of a row(Vec<String>) from table1 and a row(Vec<String>) from table2, if the first element of them are equal, then add all elements of the row from table2 except its first element to the row from table1 and add it to the results. Note that the order of results does not matter.

pythagorean

Generates sequence of unique primitive Pythagorean triples, i.e. (a,b,c) such that a² + b² = c², a and b are coprimes, and a < b. Generate in the increasing order of c.

remove_duplicate

Remove all duplicate occurences of a number inside the array. That is, if an integer appears more than once, remove some occurences of it so that it only appears once. Note that you must modify the vector using the given mutable reference instead of returning a new vector. Also, note that the order does not matter.

remove_even

Remove all even numbers inside a vector using the given mutable reference. That is, you must modify the vector using the given mutable reference instead of returning a new vector.

traverse_preorder

Traverses the tree in preorder.