Function cs220::assignments::assignment10::small_exercises::natural_join

pub fn natural_join(
    table1: Vec<Vec<String>>,
    table2: Vec<Vec<String>>,
) -> Vec<Vec<String>>

Returns the natural join of two tables using the first column as the join argument. That is, for each pair of a row(Vec<String>) from table1 and a row(Vec<String>) from table2, if the first element of them are equal, then add all elements of the row from table2 except its first element to the row from table1 and add it to the results. Note that the order of results does not matter.

Example

       table1                     table2
----------------------     ----------------------
 20230001 |    Jack         20230001 |    CS
 20231234 |    Mike         20230001 |    EE
                            20231234 |    ME

              result
-----------------------------------
 20230001 |    Jack   |     CS
 20230001 |    Jack   |     EE
 20231234 |    Mike   |     ME