Function cs220::assignments::assignment10::small_exercises::traverse_preorder

pub fn traverse_preorder<T>(root: Node<T>) -> Vec<T>

Traverses the tree in preorder.

The algorithm for preorder traversal is as follows:

Visit the root.

If the root is a leaf node, end the traverse.

If the root is a non-leaf node, traverse each subtree from the child nodes.

For example, the result of preorder traversal for the following tree

    1
   /|\
  2 3 4
 /|  /|\
5 6 7 8 9

which can be represented as

Node::NonLeaf((
    1,
    vec![
        Node::NonLeaf((2, vec![Node::Leaf(5), Node::Leaf(6)])),
        Node::Leaf(3),
        Node::NonLeaf((4, vec![Node::Leaf(7), Node::Leaf(8), Node::Leaf(9)])),
    ]
))

is 1 -> 2 -> 5 -> 6 -> 3 -> 4 -> 7 -> 8 -> 9.