Module cs220::assignments::assignment11::graph

A small graph library.

A node has a i32 value and (directed) edges to other nodes. A node does not have multiple edges to the same node. Nodes are not associated with a particular domain, and users can freely create nodes however they like. However, after a node is created, it can be added to a SubGraph, which form a subgraph of the graph of all nodes. A node can be added to multiple subgraphs. SubGraph has a method to check if the it has a cycle.

The goal of this assignment is to learn how to deal with inherently shared mutable data in Rust. Design the types and fill in the todo!()s in methods. There are several possible approaches to this problem and you may import anything from the std library accordingly.

Refer graph_grade.rs for test cases.

Structs

GraphError

Error type for graph operations.

NodeHandle

Handle to a graph node.

SubGraph

Subgraph