Trait cs220::assignments::assignment11::mock_storage::Storage

pub trait Storage {
    // Required methods
    fn upload(&self, name: &str, size: usize) -> Result<(), usize>;
    fn used(&self) -> usize;
    fn capacity(&self) -> usize;
}

Trait for storage object.

Required Methods

fn upload(&self, name: &str, size: usize) -> Result<(), usize>

Uploads a file. If a file with the same name already exists in the storage, overwrite it.

Returns Err with insufficient memory size if there is no free space to upload a file.

fn used(&self) -> usize

Returns the used memory size of the storage.

fn capacity(&self) -> usize

Returns the capacity of the storage.