Module cs220::assignments::assignment11::tv_room

TV Room Simulator.

People can come to the TV room and watch TV. There are two types of TV watchers, manager and guest.

The rule of the TV room is as follows:

Closed TV room can be opened by the manager.

Guests can enter the TV room by the manager.

Manager can leave the TV room earlier than guests.

The TV room closes when the last person left the TV room.

Both Manager and Guest have Rc<Watcher> as a field, and its reference count indicates the number of people in the TV room. When the ‘Manager’ and ‘Guest’ object is dropped, it means that the person leaves the TV room.

Consult the following documentations:

https://doc.rust-lang.org/book/ch15-04-rc.html#rct-the-reference-counted-smart-pointer

https://doc.rust-lang.org/book/ch15-05-interior-mutability.html#having-multiple-owners-of-mutable-data-by-combining-rct-and-refcellt

Refer tv_room_grade.rs for test cases.

Structs

Guest

TV Room Guest.

Manager

TV Room Manager.

TVRoom

TV Room