Trait cs220::assignments::assignment07::my_itertools::MyIterTools

pub trait MyIterTools: Iterator {
    // Provided methods
    fn my_unique(self) -> Unique<Self> 
       where Self: Sized { ... }
    fn my_chain<I: Iterator>(self, other: I) -> Chain<Self, I> 
       where Self: Sized { ... }
    fn my_enumerate(self) -> Enumerate<Self> 
       where Self: Sized { ... }
    fn my_zip<I: Iterator>(self, other: I) -> Zip<Self, I> 
       where Self: Sized { ... }
    fn my_fold<T, F>(self, init: T, f: F) -> T
       where Self: Sized,
             F: FnMut(Self::Item, T) -> T { ... }
}

My Itertools trait.

Provided Methods

fn my_unique(self) -> Unique<Self>
where Self: Sized,

Returns an iterator that iterates over the self and returns only unique elements.

fn my_chain<I: Iterator>(self, other: I) -> Chain<Self, I>
where Self: Sized,

Returns an iterator that chains self and other together.

fn my_enumerate(self) -> Enumerate<Self>
where Self: Sized,

Returns an iterator that iterates over self and enumerates each element.

fn my_zip<I: Iterator>(self, other: I) -> Zip<Self, I>
where Self: Sized,

Returns an iterator that zips self and other together.

fn my_fold<T, F>(self, init: T, f: F) -> T
where Self: Sized, F: FnMut(Self::Item, T) -> T,

Foldleft for MyIterTools